ai_chat_logger/requirements.txt
Django>=4.2
djangorestframework
drf-spectacular
openai
